Allow specifying the deadline for signup in terms of shift length
Allow finer control over the specification of the
time after shift start where signup is still allowed.

The new config field is multiplied by the shift duration,
and the result added to the start time to determine the
time when signup is closed.

The existing signup_post_minutes is just added to the time
calculated by this new feature.

The feature is useful when the signup should be allowed
not just a few minutes after shift start but for a larger part
of the shift (like for half of the shift).
With the previous option it would not make sense to allow a late signup
longer than the shortest shift of the event.

This is a follow-up to a50dd9cae09f0167c6ece831d99b975ac5b4af4a