Register page layout rework
As discussed with @MyIgel the whole register page is in dire need of gasoline and a lighter. I started to refactor it by rearranging the row & column grid (removed 2 nesting levels) and changed the order of the field to something with a little more sense. The page still needs to be moved to a proper view...