patches language files
Before we substitude mainly
"Engel" -> "Helferin"
"Himmel" -> "Helferinnenraum"
"angel" -> "volunteer"
"heaven" -> "volunteer managers"
we had to deal with the incomplete english language file.

Therefore we use `xgettext` to add create missing entries.

We saw that we could have used the `TranslationServiceProvider` [1] to
substitude the expression but decided against it for performance reason.

[1] https://github.com/MyIgel/engelsystem/commit/0a5c629deaacbb10d11673918c3579a8c867a695