Fix backup (#16)
The previous backup would backup all the databases every hour, excluding only the data of the sessions table of the main database. As it turns out, the log_entries contributes a lot of data as well, as do the databases of previous years. Unfortunately, it's not possible to ignore data based on a wildcard.
We decided to switch to a two-tiered backup: the current database is backed up hourly without sessions and log_entries, and all databases are backed up weekly. Data retention is adapted accordingly. We might still want to prune sessions from previous years eventually, but maybe that's not worth automating...